requirements.txt
pandas
requests
psycopg2-binary
sqlalchemy
python-dotenv